feat: AuthContext with routing integration
